[iOS] refactor: PlayerView 내 subview 분리
